Techsfilter includes projects that match 1 or more techs
Move helper functions to databse/utils.js
